Squashed commit of the following:
commit 54869b69664395e341306c37c5b04b2e4f8efdd8 [formerly 800f6a54c34f20c65dade8e0f376e4549ae0e0ef] [formerly 82d95ca09df7614f1e5a8cd319e02b871696928e [formerly 7822051a4ee8c58b63e8943a0a9ec4760849be42]] [formerly bdcf427352b665a466af8391dad1ad4bb5944e74] [formerly f1c4cdf2ef92cf621210b60268b76f5fc73bced0] [formerly 19ff397a71ad487ce25d65b0d6fd4ab2d97bbbb5]
Author: Ajay Panyala <ajay.panyala@gmail.com>
Date:   Thu Jul 25 14:47:32 2019 -0700

    revert Tile typedef

commit ddf723b68df748b21f5675e97ef9921c92f32470 [formerly a8cf6a40fc916551e94683c4071d83ba53d9014c] [formerly 1cca86e9609a3ebefbacc22a95dd85a355ef115d [formerly 4db17e17fcb47e91f7ea9d10020385ba1cab3d42]] [formerly f43d2d13dc64d31319d08e4f6e6f70afd845a337] [formerly f2d62feb609d9b6d904296b5e797662da6207d7e] [formerly f2fceb2046d5e196fab596ba1a8df9281f20ad3f]
Merge: 5ee4ccaa a7c498ce
Author: Erdal Mutlu <erdalmutlu.su@gmail.com>
Date:   Thu Jul 25 10:28:49 2019 -0700

    Merge branch 'develop' of https://github.com/NWChemEx-Project/TAMM into develop

commit 8e14f5573e52ab61167493e301c396e52117f36f [formerly fa23475deceb97455cde27ac155a2e0735caea72] [formerly cf68c8ce470a0e0e3be970377461b40808fc4bc5 [formerly 6334eac000131ba658bf06a92ad8c09918085ae1]] [formerly 124a8877fb11c7fcc5988e02a5855c04abb42531] [formerly 88aea1f8e507c8622de9b832d02398073a1ed586] [formerly 5ee4ccaa5b2077831c721856cca9d0b9de03931a]
Author: Erdal Mutlu <erdalmutlu.su@gmail.com>
Date:   Thu Jul 25 10:28:44 2019 -0700

    bug fixes for sparse computation


Former-commit-id: 9f5822ec20efeae2e702023d5a5fd8f169e234e0
Former-commit-id: 2b1faf9f3dfa6fbeb425ab86e153db27b3b03cae
Former-commit-id: 9da3bb2eda1cf1532924371cb9a23c5a739d615f
Former-commit-id: 949c1438e6c643fe52ee6634c7e491185b8072dd [formerly 2f8d3b1302b34783941a21995a1862c9da0738b0]
Former-commit-id: 2bc8d69612938eff512151abca4a018656e7f763

+++ src/tamm/types.hpp
@@ -45,7 +45,7 @@ using TAMM_SIZE = uint64_t;
 using Index         = uint32_t;
 using IndexVector   = std::vector<Index>;
 using IndexIterator = std::vector<Index>::const_iterator;
-using Tile          = size_t;
+using Tile          = uint32_t;
 // DAG related Hash
 using HashData      = uint64_t;
 using StringLabelVec    = std::vector<std::string>;
